Minor clean up.
